Fix runtime initialization issue
During initialization, MonoTargetRuntime loads packages from .pc files.
The process of parsing .pc files uses SystemAssemblyService to find out
which frameworks do packages belong to. This query sometimes fails because
frameworks are loaded *after* runtimes are initialized, when the runtime
rises the TargetRuntime.Initialized event. The solution is to add a new event,
TargetRuntime.FrameworksInitialized, which is raised when frameworks are
created but before the .pc files are parsed. SystemAssemblyService now
subscribes TargetRuntime.FrameworksInitialized instead of
TargetRuntime.Initialized, so that frameworks are properly loaded before
.pc file processing.